requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-oidc==1.4.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.14.0
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
